+++ packages/www/src/documentation/components/content/demos/index.ts
@@ -26,3 +26,4 @@
 
 export * from './IconList'
 export * from './DateTimeFormat'
+export * from './CalendarPropTable'